some small corrects

+++ cmd/main.go
@@ -17,14 +17,20 @@ func main() {
 
 	// Точка входа программы:
 	// go run .\cmd\main.go [команда]...
-	// add [Заголовок]
+	// add ["Заголовок"]  		*в кавычках ("")
 	// list
-	// update [id] [Новый заголовок]
+	// update [id] ["Новый заголовок"]
 	// delete [id]
-
+	if len(os.Args)<2{
+		fmt.Print(storage.СommandList())
+		return
+	}
 	switch  os.Args[1]{
 	case "add":
-
+		if len(os.Args) < 3 {
+			fmt.Print("Неверный ввод команды add!\n", storage.СommandList())
+			return
+		}
 		lastID, err := storage.FindLastId(filename)
 		if err != nil{
 			log.Fatal("Ошибка нахождения последнего айди:",err)
@@ -33,7 +39,7 @@ func main() {
 		createdTime := time.Now()
 		task := task.AddTask(lastID + 1, os.Args[2], true, createdTime, nil)
 	
-		log.Println("Task added:", task)
+		log.Print("Task added:\n", storage.ListTask(*task))
 	
 		err = storage.SaveTasksToFile(filename, task)
 		if err != nil {
@@ -50,6 +56,10 @@ func main() {
 		
 	case "update":
 		// Call UpdateTask function
+		if len(os.Args) < 4 {
+			fmt.Print("Неверный ввод команды update!\n", storage.СommandList())
+			return
+		}
 		err := storage.UpdateTask(filename)
 		if err != nil{
 			log.Fatal("Ошибка записи в файл Update:",err)
@@ -57,7 +67,10 @@ func main() {
 		
 	case "delete":
 		// Call DeleteTask function
-		// Запись измененной мапы в файл
+		if len(os.Args) < 3 {
+			fmt.Print("Неверный ввод команды delete!\n", storage.СommandList())
+			return
+		}
 		err := storage.DeleteTask(filename)
 		if err != nil{
 			log.Fatal("Error saving tasks to file:", err)
@@ -68,3 +81,4 @@ func main() {
 		
 	}
 }
+
